apis/app/analytics/functions/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
passlib[bcrypt]==1.7.4  # Includes bcrypt support
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.10
bcrypt==4.1.3
gunicorn==21.2.0
cryptography==42.0.7
openai==1.55.3
faiss-cpu==1.8.0
langgraph==0.0.62
langchain
langchain-core
langchain-openai==0.1.9
langchain-community
langchain-groq
langsmith==0.1.113
tenacity!=8.4.0,<9.0.0
pydantic==1.10.13  
sqlmodel==0.0.16
SQLAlchemy
